pyproject.toml
[tool.black]
line-length = 79
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.github
    | \.hg
    | \.mypy_cache
    | \.pytest_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 79
multi_line_output = 3

[tool.poetry]
name = "builtwithdjango"
version = "0.1.0"
description = "Source code for the builtwithdjango.com website."
authors = ["Rasul Kireev <me@rasulkireev.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.4"
django-autoslug = "^1.9.7"
django-environ = "^0.4.5"
django-taggit = "^1.2.0"
djangorestframework = "^3.11.0"
gunicorn = "^20.0.4"
requests = "^2.23.0"
sentry_sdk = "^0.14.3"
django-component = "^0.1.6"
Pillow = ">=8.0.0"
Markdown = "^3.3.4"
django-model-utils = "^4.1.1"
django-extensions = "^3.1.3"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
black = "^21.5b2"
pylint = "^2.5.3"
pre-commit = "^2.6.0"
pylint-django = "^2.2.0"
flake8 = "^3.9.2"
djlint = "^0.2.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
